% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beer.R
\name{createBeer}
\alias{createBeer}
\title{Add a new beer}
\usage{
createBeer(name, styleId, description, abv, ibu, glasswareId, srmId,
  availableId, isRetired, isOrganic, beerVariationId, year, foodPairings,
  servingTemperature, originalGravity, brewery, label,
  nutritionServingSize, calories, fat, saturatedFat, transFat, cholesterol,
  sodium, carbohydrates, dietaryFiber, sugars, protein)
}
\arguments{
\item{name}{Required Name of the beer}

\item{styleId}{Required ID corresponding to the appropriate style}

\item{description}{A little bit about the beer}

\item{abv}{ABV percentage of the beer}

\item{ibu}{Measured IBUs of the beer}

\item{glasswareId}{ID corresponding to the appropriate glassware}

\item{srmId}{ID corresponding to the appropriate SRM}

\item{availableId}{ID corresponding to the appropriate beer availability setting}

\item{isRetired}{Whether or not the beer is produced or not Y N Default}

\item{isOrganic}{Whether or not the beer is certified organic or not Y N Default}

\item{beerVariationId}{ID of an existing beer that this beer is a variation of}

\item{year}{Vintage of the beer, if applicable}

\item{foodPairings}{Information about what foods should be paired with the beer}

\item{servingTemperature}{Recommended serving temperature of the beer}

\item{originalGravity}{Measured original gravity of the beer}

\item{brewery}{Comma separated list of existing brewery IDs to add this beer to}

\item{label}{Base64 encoded image to be assigned as the beer label}

\item{nutritionServingSize}{Displays the serving size that the nutritional information based on.}

\item{calories}{Total calories in beer based on serving size. Serving size is Required if this is set.}

\item{fat}{Total Fat in beer based on serving size. Serving size is Required if this is set.}

\item{saturatedFat}{Total Saturated Fat in beer based on serving size. Serving size is Required if this is set.}

\item{transFat}{Total Trans Fat in beer based on serving size. Serving size is Required if this is set.}

\item{cholesterol}{Total Cholesterol in beer based on serving size. Serving size is Required if this is set.}

\item{sodium}{Total Sodium in beer based on serving size. Serving size is Required if this is set.}

\item{carbohydrates}{Total Carbohydrates in beer based on serving size. Serving size is Required if this is set.}

\item{dietaryFiber}{Total Dietary Fiber in beer based on serving size. Serving size is Required if this is set.}

\item{sugars}{Total Sugars in beer based on serving size. Serving size is Required if this is set.}

\item{protein}{Total Protein in beer based on serving size. Serving size is Required if this is set.}
}
\value{
none
}
\description{
Add a beer.  Beer will be added with status of new_unverified.
}
\concept{Beer}
