% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerAdjunct.R
\name{createBeerAdjunct}
\alias{createBeerAdjunct}
\title{Add an Adjunct}
\usage{
createBeerAdjunct(beerId, adjunctId)
}
\arguments{
\item{beerId}{The beerId}

\item{adjunctId}{Required ID of an existing adjunct}
}
\value{
none
}
\description{
Adds an adjunct to an existing beer.
}
\concept{BeerAdjunct}
