% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerBrewery.R
\name{createBeerBrewery}
\alias{createBeerBrewery}
\title{Add a Brewery}
\usage{
createBeerBrewery(beerId, breweryId, locationId)
}
\arguments{
\item{beerId}{The beerId}

\item{breweryId}{Required ID of an existing brewery}

\item{locationId}{Specific location of a brewery to add a beer to. If not set, will be added to all locations.}
}
\value{
none
}
\description{
Adds an existing brewery to an existing beer.
}
\concept{BeerBrewery}
