% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerFermentable.R
\name{createBeerFermentable}
\alias{createBeerFermentable}
\title{Add a Fermentable}
\usage{
createBeerFermentable(beerId, fermentableId)
}
\arguments{
\item{beerId}{The beerId}

\item{fermentableId}{Required ID of an existing fermentable}
}
\value{
none
}
\description{
Adds a fermentable to an existing beer.
}
\concept{BeerFermentable}
