% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerHop.R
\name{createBeerHop}
\alias{createBeerHop}
\title{Add a Hop}
\usage{
createBeerHop(beerId, hopId)
}
\arguments{
\item{beerId}{The beerId}

\item{hopId}{Required ID of an existing hop}
}
\value{
none
}
\description{
Adds a hop to an existing beer.
}
\concept{BeerHop}
