% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerSocialaccount.R
\name{createBeerSocialaccount}
\alias{createBeerSocialaccount}
\title{Add a Social Account}
\usage{
createBeerSocialaccount(beerId, socialMediaId, handle)
}
\arguments{
\item{beerId}{The beerId}

\item{socialMediaId}{Required ID of an existing social media site}

\item{handle}{Required Handle, or identifying marker, for the specific social media site}
}
\value{
none
}
\description{
Adds a social account to an existing beer.
}
\concept{BeerSocialaccount}
