% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerUpc.R
\name{createBeerUpc}
\alias{createBeerUpc}
\title{Add a UPC Code}
\usage{
createBeerUpc(beerId, upcCode, fluidSizeId)
}
\arguments{
\item{beerId}{The beerId}

\item{upcCode}{Required UPC code to assign to the beer}

\item{fluidSizeId}{ID of an existing container size}
}
\value{
none
}
\description{
Adds a UPC code to an existing beer.
}
\concept{BeerUpc}
