% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerYeast.R
\name{createBeerYeast}
\alias{createBeerYeast}
\title{Add a Yeast}
\usage{
createBeerYeast(beerId, yeastId)
}
\arguments{
\item{beerId}{The beerId}

\item{yeastId}{Required ID of an existing yeast}
}
\value{
none
}
\description{
Adds a yeast to an existing beer.
}
\concept{BeerYeast}
