% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brewery.R
\name{createBrewery}
\alias{createBrewery}
\title{Add a new brewery}
\usage{
createBrewery(name, description, website, established, mailingListUrl,
  isOrganic, image)
}
\arguments{
\item{name}{Required Name of the brewery}

\item{description}{A little bit about the brewery}

\item{website}{Url of the breweries website}

\item{established}{Year brewery was established. Format YYYY}

\item{mailingListUrl}{Url of the breweries mailing list}

\item{isOrganic}{Y N Default}

\item{image}{Base64 encoded image}
}
\value{
none
}
\description{
Add a brewery.  Brewery will be added with status of new_unverified.
}
\concept{Brewery}
