% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryAlternatename.R
\name{createBreweryAlternatename}
\alias{createBreweryAlternatename}
\title{Add a Alternate Name}
\usage{
createBreweryAlternatename(breweryId, name)
}
\arguments{
\item{breweryId}{The breweryId}

\item{name}{Required Alternate Name of the brewery}
}
\value{
none
}
\description{
Add a Alternate name to a brewery.
}
\concept{BreweryAlternatename}
