% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryGuild.R
\name{createBreweryGuild}
\alias{createBreweryGuild}
\title{Add a Guild to brewery}
\usage{
createBreweryGuild(breweryId, guildId, discount)
}
\arguments{
\item{breweryId}{The breweryId}

\item{guildId}{Required Id of the Guild}

\item{discount}{If the brewery offers a discount for members of this guild, put it here.}
}
\value{
none
}
\description{
Associates a guild to a specific brewery.
}
\concept{BreweryGuild}
