% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryLocation.R
\name{createBreweryLocation}
\alias{createBreweryLocation}
\title{Add a new location}
\usage{
createBreweryLocation(breweryId, name, streetAddress, extendedAddress,
  locality, region, postalCode, phone, website,
  hoursOfOperationExplicitString, hoursOfOperationNotes, tourInfo,
  timezoneId, latitude, longitude, isPrimary, inPlanning, isClosed,
  openToPublic, locationType, countryIsoCode)
}
\arguments{
\item{breweryId}{The breweryId}

\item{name}{Nickname for the location}

\item{streetAddress}{Street address}

\item{extendedAddress}{Extended address, such as suite or apartment number}

\item{locality}{Locality, or city, where the location is at}

\item{region}{Region, also known as state or province}

\item{postalCode}{Postal code}

\item{phone}{Phone number for the location}

\item{website}{Location-specific website, if different from the brewery}

\item{hoursOfOperationExplicitString}{Explict breakdown of the hours of operation. Should be in format "ddd-hh:mm(am/pm)-hh:mm(am/pm)". Multiple days can be passed with the given format, separated by commas. Example: 'mon-8:00am-10:00pm,tue-9:00am-1:00pm' is valid.}

\item{hoursOfOperationNotes}{Additional, non-time-related notes about the hours of operation.}

\item{tourInfo}{Brewery tour information for visitors of the location}

\item{timezoneId}{Timezone ID for the location. This value is automatically updated based on the latitude and longitude of the location, but can be passed as well. Example: America/New_York}

\item{latitude}{Latitude for the location}

\item{longitude}{Longitude for the location}

\item{isPrimary}{Whether or not this is the primary location for the brewery Y N Default}

\item{inPlanning}{Whether or not this location is in planning Y N Default}

\item{isClosed}{Whether or not this location is currently closed Y N Default}

\item{openToPublic}{Whether or not this location is open to the public Y N Default}

\item{locationType}{Type of location, such as micro brewery or brew pub micro macro nano prewpub production office tasting restaurant cidery meadery}

\item{countryIsoCode}{Required Country that the location is located in}
}
\value{
none
}
\description{
Add a new Location to a brewery.  Location will be added with status of new_unverified.
}
\concept{BreweryLocation}
