% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrewerySocialaccount.R
\name{createBrewerySocialaccount}
\alias{createBrewerySocialaccount}
\title{Add a Social Account}
\usage{
createBrewerySocialaccount(breweryId, socialmediaId, handle)
}
\arguments{
\item{breweryId}{The breweryId}

\item{socialmediaId}{Required ID of an existing social media site}

\item{handle}{Required Handle, or identifying marker, for the specific social media site}
}
\value{
none
}
\description{
Adds a social account to an existing brewery.
}
\concept{BrewerySocialaccount}
