% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convertid.R
\name{createConvertid}
\alias{createConvertid}
\title{Convert IDs}
\usage{
createConvertid(type, ids)
}
\arguments{
\item{type}{Required Type of entity to convert the IDs for brewery beer}

\item{ids}{Required Comma delimited list of BreweryDB API v1 IDs}
}
\value{
none
}
\description{
Takes a type (brewery or beer), and a comma delimited list of v1 id's and returns the mapping to corresponding v2 ids.
}
\concept{Convertid}
