% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guild.R
\name{createGuild}
\alias{createGuild}
\title{Add an guild}
\usage{
createGuild(name, description, established, website, image)
}
\arguments{
\item{name}{Required Name of the guild}

\item{description}{Description of the guild}

\item{established}{Year the guild was established}

\item{website}{Website of the guild}

\item{image}{An image or logo for the guild, base64 encoded source.}
}
\value{
none
}
\description{
Adds an guild to the system.
}
\concept{Guild}
