% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuildSocialaccount.R
\name{createGuildSocialaccount}
\alias{createGuildSocialaccount}
\title{Add a Social Account}
\usage{
createGuildSocialaccount(guildId, socialmediaId, handle)
}
\arguments{
\item{guildId}{The guildId}

\item{socialmediaId}{Required ID of an existing social media site}

\item{handle}{Required Handle, or identifying marker, for the specific social media site}
}
\value{
none
}
\description{
Adds a social account to an existing guild.
}
\concept{GuildSocialaccount}
