% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beer.R
\name{deleteBeer}
\alias{deleteBeer}
\title{Delete a beer}
\usage{
deleteBeer(beerId)
}
\arguments{
\item{beerId}{The beerId}
}
\value{
none
}
\description{
Delete a beer  Beer will have a status of delete_pending after.
}
\concept{Beer}
