% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerHop.R
\name{deleteBeerHop}
\alias{deleteBeerHop}
\title{Remove a Hop From Beer}
\usage{
deleteBeerHop(hopId, beerId)
}
\arguments{
\item{hopId}{The hopId}

\item{beerId}{The beerId}
}
\value{
none
}
\description{
Removes the given hop from the beer.
}
\concept{BeerHop}
