% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerSocialaccount.R
\name{deleteBeerSocialaccount}
\alias{deleteBeerSocialaccount}
\title{Delete a Social Account From Beer}
\usage{
deleteBeerSocialaccount(socialaccountId, beerId)
}
\arguments{
\item{socialaccountId}{The socialaccountId}

\item{beerId}{The beerId}
}
\value{
none
}
\description{
Deletes a social account for the beer.
}
\concept{BeerSocialaccount}
