% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerYeast.R
\name{deleteBeerYeast}
\alias{deleteBeerYeast}
\title{Remove a Yeast From Beer}
\usage{
deleteBeerYeast(yeastId, beerId)
}
\arguments{
\item{yeastId}{The yeastId}

\item{beerId}{The beerId}
}
\value{
none
}
\description{
Removes the given yeast from the beer.
}
\concept{BeerYeast}
