% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brewery.R
\name{deleteBrewery}
\alias{deleteBrewery}
\title{Delete a new brewery}
\usage{
deleteBrewery()
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Delete a brewery.  Brewery will be added with status of delete_pending.
}
\concept{Brewery}
