% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryAlternatename.R
\name{deleteBreweryAlternatename}
\alias{deleteBreweryAlternatename}
\title{Delete a brewery alternate name}
\usage{
deleteBreweryAlternatename(alternatenameId, breweryId)
}
\arguments{
\item{alternatenameId}{The alternatenameId}

\item{breweryId}{The breweryId}
}
\value{
none
}
\description{
Delete a breweries alternate Name.
}
\concept{BreweryAlternatename}
