% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryGuild.R
\name{deleteBreweryGuild}
\alias{deleteBreweryGuild}
\title{Delete brewery guild association}
\usage{
deleteBreweryGuild(guildId, breweryId)
}
\arguments{
\item{guildId}{The guildId}

\item{breweryId}{The breweryId}
}
\value{
none
}
\description{
Delete a guild from the brewery.
}
\concept{BreweryGuild}
