% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrewerySocialaccount.R
\name{deleteBrewerySocialaccount}
\alias{deleteBrewerySocialaccount}
\title{Delete a Social Account From Brewery}
\usage{
deleteBrewerySocialaccount(socialaccountId, breweryId)
}
\arguments{
\item{socialaccountId}{The socialaccountId}

\item{breweryId}{The breweryId}
}
\value{
none
}
\description{
Deletes a social account for the brewery.
}
\concept{BrewerySocialaccount}
