% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuildSocialaccount.R
\name{deleteGuildSocialaccount}
\alias{deleteGuildSocialaccount}
\title{Deletes a Social Account For Guild}
\usage{
deleteGuildSocialaccount(socialaccountId, guildId, socialmediaId, handle)
}
\arguments{
\item{socialaccountId}{The socialaccountId}

\item{guildId}{The guildId}

\item{socialmediaId}{Required ID of an existing social media site}

\item{handle}{Required Handle, or identifying marker, for the specific social media site}
}
\value{
none
}
\description{
Deletes a social account for the guild.
}
\concept{GuildSocialaccount}
