% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location.R
\name{deleteLocation}
\alias{deleteLocation}
\title{Delete an Location}
\usage{
deleteLocation(locationId)
}
\arguments{
\item{locationId}{The locationId}
}
\value{
none
}
\description{
Deletes an existing location
}
\concept{Location}
