% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heartbeat.R
\name{EveryHeartbeat}
\alias{EveryHeartbeat}
\title{Get Heartbeat}
\usage{
EveryHeartbeat()
}
\arguments{
\item{none}{}
}
\value{
none
}
\description{
Ensures the API is up and listening.  Will take any list of arguments, and return them back along with a timestamp.
}
\concept{Heartbeat}
