% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjunct.R
\name{getAdjunct}
\alias{getAdjunct}
\title{Get An Adjunct}
\usage{
getAdjunct(adjunctId)
}
\arguments{
\item{adjunctId}{The adjunctId}
}
\value{
none
}
\description{
Gets a specific adjunct by ID.
}
\concept{Adjunct}
