% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerSocialaccount.R
\name{getBeerSocialaccount}
\alias{getBeerSocialaccount}
\title{Get Specific Social Account For Beer}
\usage{
getBeerSocialaccount(socialaccountId, beerId)
}
\arguments{
\item{socialaccountId}{The socialaccountId}

\item{beerId}{The beerId}
}
\value{
none
}
\description{
Gets a specific social account for a specific beer.
}
\concept{BeerSocialaccount}
