% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brewery.R
\name{getBrewery}
\alias{getBrewery}
\title{Get a single brewery}
\usage{
getBrewery(withSocialAccounts, withGuilds, withLocations,
  withAlternateNames)
}
\arguments{
\item{withSocialAccounts}{Get brewery results with social account information included. Y N Default}

\item{withGuilds}{Get brewery results with guild information included. Y N Default}

\item{withLocations}{Get brewery results with location information included. Y N Default}

\item{withAlternateNames}{Get brewery results with alternate name information included. Y N Default}
}
\value{
none
}
\description{
Gets a single brewery
}
\concept{Brewery}
