% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrewerySocialaccount.R
\name{getBrewerySocialaccount}
\alias{getBrewerySocialaccount}
\title{Get Specific Social Account For Brewery}
\usage{
getBrewerySocialaccount(socialaccountId, breweryId)
}
\arguments{
\item{socialaccountId}{The socialaccountId}

\item{breweryId}{The breweryId}
}
\value{
none
}
\description{
Gets a specific social account for a specific brewery.
}
\concept{BrewerySocialaccount}
