% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Category.R
\name{getCategory}
\alias{getCategory}
\title{Get a single category}
\usage{
getCategory(categoryId)
}
\arguments{
\item{categoryId}{The categoryId}
}
\value{
none
}
\description{
Gets a single Category
}
\concept{Category}
