% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjunct.R
\name{getEveryAdjunct}
\alias{getEveryAdjunct}
\title{Get List of Adjuncts}
\usage{
getEveryAdjunct(p)
}
\arguments{
\item{p}{Page Number}
}
\value{
id The unique id of the adjunct.

name The name of the adjunct.

description The description of the adjunct.

category This value will always be set to "misc".

categoryDisplay This value will always be set to "Miscellaneous".
}
\description{
Gets a list of all adjuncts. Results are paginated.
}
\concept{Adjunct}
