% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beer.R
\name{getEveryBeer}
\alias{getEveryBeer}
\title{Get All Beers}
\usage{
getEveryBeer()
}
\arguments{
\item{p}{Page Number}

\item{ids}{ID's of the beers to return, comma separated. Max 10.}

\item{name}{Name of a beer.}

\item{abv}{ABV for a beer. Premium users may use advanced filtering. "+10" will return everything above 10%, "-10" will return everything under 10%, "8,10" will return everything between 8% and 10% inclusive.}

\item{ibu}{IBUs for a beer. Premium users may use advanced filtering. "+50" will return everything above 50 IBUs, "-50" will return everything less than 50 IBUs, "30,50" will return everything between 30 and 50 IBUs inclusive.}

\item{glasswareId}{ID for glassware}

\item{srmId}{ID for SRM}

\item{availableId}{ID for availability}

\item{styleId}{ID for style}

\item{isOrganic}{Whether the beer is certified organic or not Y N}

\item{hasLabels}{Whether or not the beer has a label Y N}

\item{year}{Year vintage of the beer. Format YYYY}

\item{since}{Returns everything that has been updated since that date. Max 30 days. In UNIX timestamp format.}

\item{status}{Status of the brewery in the API}

\item{order}{How the results should be ordered name Default description abv ibu glasswareId srmId availableId styleId isOrganic status createDate updateDate random}

\item{sort}{How the results should be sorted. ASC Default DESC}

\item{randomCount}{If the order parameter is set to random, this option specifies how many random beers to return. It has a max value of 10}

\item{withBreweries}{Get beer results with brewery information included. Y N Default}

\item{withSocialAccounts}{Get beer results with social account information included. Y N Default}

\item{withIngredients}{Get beer results with ingredients information included. Y N Default}
}
\value{
id The unique id of the beer.

name The name of the beer.

nameDisplay Display name of the beer.

description The description of the beer.

foodPairings A free text field containing any food pairing information for the beer.

originalGravity The original gravity of the beer.

abv The alcohol by volume of the beer (expressed as a percentage).

ibu The IBU (international bittering unit) value is a measure of how bitter a beer is. The higher the number, the more bitter the beer.

isRetired Displays if the beer is no longer produced. ('Y' or 'N')

glasswareId The id corresponding to the glass object that is assigned to this beer. If this exists, then so will the glass object. See the glass object for more information.

glass Contains the details about the assigned glass (id, name). The name of the glass is the type of glass in which the beer is best served.

styleId The id corresponding to the style object that is assigned to the beer. See the style object for more information.

style The style object contains details about the assigned style: id, categoryId, category (id, name), name, description, ibuMin, ibuMax, abvMax, srmMin, srmMax, ogMin, fgMin, fgMax).

isOrganic Whether or not the beer is certified organic.

labels If this object is set then labels exist and it will contain items icon, medium, and large that are URLs to the images.

servingTemperature The key that corresponds to the serving temperature information. See the servingTemperatureDisplay field for the full details.

servingTemperatureDisplay The serving temperature information. If the servingTemperature was "cool" then this field would be like "Cool - (8-12C/45-54F)"

status The status key for the object

statusDisplay The display string corresponding to the status.

availableId The id corresponding to the assigned availability object.

available The object that provides details for the assigned availability. It contain: id, name, and description.

beerVariationId If this beer is a variation of another beer, then the beerVariationId will be set to the id of the "source" beer. For example, if "My IPA" has id abc123 and you look up "Oak Aged My IPA" (id def456) then the beerVariationId of "Oak Aged My IPA" would be abc123.

beerVariation If this object will be set if there is a beerVariationId set. This is an instance of a beer object that has all the information for the "source" beer if the beer is a variation.

year The year field is for vintages of a beer.

nutritionServingSize Displays the serving size that the nutritional information based on.

nutritionEstimated Displays if the beer has estimated nutritional values. ('Y' or 'N')

calories Total calories in beer based on Serving Size

fat Total Fat in beer based on Serving Size

saturatedFat Total Saturated Fat in beer based on Serving Size

transFat Total Trans Fat in beer based on Serving Size

cholesterol Total Cholesterol in beer based on Serving Size

sodium Total Sodium in beer based on Serving Size

carbohydrates Total Carbohydrates in beer based on Serving Size

dietaryFiber Total Dietary Fiber in beer based on Serving Size

sugars Total Sugars in beer based on Serving Size

protein Total Protein in beer based on Serving Size
}
\description{
Gets a listing of all beers.  Results will be paginated.
}
\concept{Beer}
