% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerBrewery.R
\name{getEveryBeerBrewery}
\alias{getEveryBeerBrewery}
\title{Get All Breweries For Beer}
\usage{
getEveryBeerBrewery(beerId)
}
\arguments{
\item{beerId}{The beerId}
}
\value{
none
}
\description{
Gets a listing of all breweries that brew the specific beer.
}
\concept{BeerBrewery}
