% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerEvent.R
\name{getEveryBeerEvent}
\alias{getEveryBeerEvent}
\title{Gets a listing of all events that a beer is at or has won awards at.}
\usage{
getEveryBeerEvent(beerId, onlyWinners)
}
\arguments{
\item{beerId}{The beerId}

\item{onlyWinners}{Only return events where the beer has won an award Y N}
}
\value{
none
}
\description{
Gets a listing of all events that a beer is at or has won awards at.
}
\concept{BeerEvent}
