% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerSocialaccount.R
\name{getEveryBeerSocialaccount}
\alias{getEveryBeerSocialaccount}
\title{Get All Social Accounts For Beer}
\usage{
getEveryBeerSocialaccount(beerId)
}
\arguments{
\item{beerId}{The beerId}
}
\value{
none
}
\description{
Gets a listing of all social accounts for a specific beer.
}
\concept{BeerSocialaccount}
