% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brewery.R
\name{getEveryBrewery}
\alias{getEveryBrewery}
\title{Get All Breweries}
\usage{
getEveryBrewery(p, name, ids, established, isOrganic, hasImages, since,
  status, order, sort, randomCount, withSocialAccounts, withGuilds,
  withLocations, withAlternateNames)
}
\arguments{
\item{p}{Page Number}

\item{name}{Name of a brewery.}

\item{ids}{ID's of the breweries to return, comma separated. Max 10.}

\item{established}{Year a brewery was established. Format YYYY}

\item{isOrganic}{Is the brewery an Organic brewery Y N Default}

\item{hasImages}{Whether the brewery has an image or not Y N}

\item{since}{Returns everything that has been updated since that date. Max 30 days. In UNIX timestamp format.}

\item{status}{Status of the brewery in the API}

\item{order}{How the results should be ordered name Default description website established mailingListUrl isOrganic status createDate updateDate random}

\item{sort}{How the results should be sorted. ASC Default DESC}

\item{randomCount}{If the order parameter is set to random, this option specifies how many random breweries to return. It has a max value of 10}

\item{withSocialAccounts}{Get brewery results with social account information included. Y N Default}

\item{withGuilds}{Get brewery results with guild information included. Y N Default}

\item{withLocations}{Get brewery results with location information included. Y N Default}

\item{withAlternateNames}{Get brewery results with alternate name information included. Y N Default}
}
\value{
id The unique id of the brewery.

name The name of the brewery.

nameShortDisplay The short name of the brewery.

isVerified Wheather or not the brewery has been claimed.

description The description of the brewery.

website The URL to the brewery's website

established The year that the brewery was established.

mailingListUrl The url to the brewery's mailing list.

isOrganic Whether or not the brewery is certified organic.

images The images field will contain the three image sizes for the brewery logos (icon, medium, and large).
}
\description{
Gets a listing of all breweries.  Results will be paginated.  One of the following attributes must be set:  name, established
}
\concept{Brewery}
