% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryAlternatename.R
\name{getEveryBreweryAlternatename}
\alias{getEveryBreweryAlternatename}
\title{Get All Breweries Alternate Names}
\usage{
getEveryBreweryAlternatename(breweryId)
}
\arguments{
\item{breweryId}{The breweryId}
}
\value{
none
}
\description{
Gets a listing of all breweries alternate names.
}
\concept{BreweryAlternatename}
