% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryEvent.R
\name{getEveryBreweryEvent}
\alias{getEveryBreweryEvent}
\title{Gets all the events a brewery is present at.}
\usage{
getEveryBreweryEvent(breweryId, onlyWinners)
}
\arguments{
\item{breweryId}{The breweryId}

\item{onlyWinners}{Display only events where the brewery has won an award. Y N}
}
\value{
none
}
\description{
Gets all the events a brewery is present at.
}
\concept{BreweryEvent}
