% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryGuild.R
\name{getEveryBreweryGuild}
\alias{getEveryBreweryGuild}
\title{Get All Guilds}
\usage{
getEveryBreweryGuild(breweryId, p)
}
\arguments{
\item{breweryId}{The breweryId}

\item{p}{Page Number}
}
\value{
none
}
\description{
Gets all guilds that a brewery belongs to.  Results will be paginated
}
\concept{BreweryGuild}
