% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreweryLocation.R
\name{getEveryBreweryLocation}
\alias{getEveryBreweryLocation}
\title{Get locations for a brewery}
\usage{
getEveryBreweryLocation(breweryId)
}
\arguments{
\item{breweryId}{The breweryId}
}
\value{
none
}
\description{
Gets a listing of all locations for the specified brewery.  Results will be paginated.
}
\concept{BreweryLocation}
