% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrewerySocialaccount.R
\name{getEveryBrewerySocialaccount}
\alias{getEveryBrewerySocialaccount}
\title{Get All Social Accounts For Brewery}
\usage{
getEveryBrewerySocialaccount(breweryId)
}
\arguments{
\item{breweryId}{The breweryId}
}
\value{
none
}
\description{
Gets a listing of all social accounts for a specific brewery.
}
\concept{BrewerySocialaccount}
