% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Category.R
\name{getEveryCategory}
\alias{getEveryCategory}
\title{Get All Categories}
\usage{
getEveryCategory()
}
\arguments{
\item{none}{}
}
\value{
id The unique id of the category.

name The name of the category.
}
\description{
Gets a listing of All categories
}
\concept{Category}
