% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Change.R
\name{getEveryChange}
\alias{getEveryChange}
\title{Get All Changes}
\usage{
getEveryChange(p, attributeName, attributeId, since)
}
\arguments{
\item{p}{Page Number}

\item{attributeName}{Name of an attribute in BreweryDB. Multiple attributeNames can be separated by a comma. beer brewery location guild event}

\item{attributeId}{ID of an attribute in BreweryDB. Will only be recognized if a single attributeName parameter is also passed.}

\item{since}{UNIX timestamp of a date in which to get changes since. This can be a max of 30 days in the past.}
}
\value{
attributeName Attribute name of the affected attribute. Either "beer", "brewery", "location", "guild" or "event".

action Action taken on that attribute. Either "edit", "insert", or "delete".

attribute Data structure related to the type of attribute that changes. If a beer was changed, you would get the basic beer structure. If a brewery was changed, you would get the basic brewery structure.

subAttributeName Name of the sub attribute that was changed. Only set in the case of relationships.

subAction Action taken on the sub attribute. Either "edit", "insert", or "delete".

subAttribute Data structure related to the type of attribute that changed.
}
\description{
Gets a listing of all changes to all attributes that have happened in the last 30 days.  Results will be paginated with 50 results per page.
}
\concept{Change}
