% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fermentable.R
\name{getEveryFermentable}
\alias{getEveryFermentable}
\title{Get List of Fermentables}
\usage{
getEveryFermentable(p)
}
\arguments{
\item{p}{Page Number}
}
\value{
id The unique id of the fermentable.

name The name of the fermentable.

description The description of the fermentable.

countryOfOrigin The two-letter ISO country code. For the complete country name and information, see the country field.

srmId The ID of the corresponding SRM object. See the srm field for more information like hex color.

srmPrecise The precise SRM value of the fermentable. Whereas the srmId is intended for things like providing the color, the srmPrecise value is ideal for things like calculations in recipes. Expressed as a percentage.

moistureContent The average moisture content of the grain, expressed as a percent of weight.

coarseFineDifference For grains, the difference between the coarse mill yield and a fine powder milling of the grain.

diastaticPower A measure of how much starch converting enzyme is in the grain. A higher number (common in pale malts) indicates the presence of more enzymes for starch conversion. Expressed as a percentage.

dryYield Equivalent to the potential field, but expressed as a percentage of total weight converted.

potential The potential yield of the ingredient, expressed in specific gravity units.

protein The percentage of the fermentable that is protein.

solubleNitrogenRatio The ratio of soluble nitrogen to total nitrogen. Expressed as a percentage. The soluble nitrogen ratio is an important indicator of malt modification. The higher the number, the more highly modified the malt.

maxInBatch The maximum recommended percentage of the mash makeup that this fermentable can be.

requiresMashing Whether or not the fermentable should be mashed.

category This value will always be set to "malt".

categoryDisplay This value will always be set to "Malts, Grains, & Fermentables".

srm The is an object that provides the details about the assigned srmId.

country This is an object that provides more information about the country matching the hop's countryOfOrigin value.

characteristics The list of characteristics / qualities of the fermentable.
}
\description{
Gets a list of all fermentables.  Results are paginated.
}
\concept{Fermentable}
