% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glass.R
\name{getEveryGlass}
\alias{getEveryGlass}
\title{Get List of Glasses}
\usage{
getEveryGlass(p)
}
\arguments{
\item{p}{Page Number}
}
\value{
id The unique id of the glass.

name The name of the glass.
}
\description{
Gets a list of all glassware.  Results are paginated.
}
\concept{Glass}
