% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guild.R
\name{getEveryGuild}
\alias{getEveryGuild}
\title{Get All Guilds}
\usage{
getEveryGuild(p, ids, name, established, since, status, hasImages, order,
  sort)
}
\arguments{
\item{p}{Page Number}

\item{ids}{ID's of the guilds to return, comma separated. Max 10.}

\item{name}{Name of a guild}

\item{established}{Year the guild was established}

\item{since}{Returns everything that has been updated since that date. Max 30 days. In UNIX timestamp format.}

\item{status}{Status of the guild in the API}

\item{hasImages}{Whether the guild has an image or not Y N}

\item{order}{How the results should be ordered name Default description established status createDate updateDate}

\item{sort}{How the results should be sorted. ASC Default DESC}
}
\value{
id The unique id of the guild.

name The name of the guild.

description The description of the guild.

website The URL to the website for the guild.

images The images object contains URLs to the logo of the guild. This will contain values for icon, medium, and large.

established The year that the guild was established.
}
\description{
Gets a listing of all guilds.  One of the following attributes must be set:  name
}
\concept{Guild}
