% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuildSocialaccount.R
\name{getEveryGuildSocialaccount}
\alias{getEveryGuildSocialaccount}
\title{Get All Social Accounts For Guild}
\usage{
getEveryGuildSocialaccount(guildId)
}
\arguments{
\item{guildId}{The guildId}
}
\value{
none
}
\description{
Gets a listing of all social accounts for a specific guild.
}
\concept{GuildSocialaccount}
