% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hop.R
\name{getEveryHop}
\alias{getEveryHop}
\title{Get List of Hops}
\usage{
getEveryHop(p)
}
\arguments{
\item{p}{Page Number}
}
\value{
id The unique id of the hop.

name The name of the hop.

description The description of the hop.

countryOfOrigin The two-letter ISO country code. For the complete country name and information, see the country field.

alphaAcidMin The minimum value for the typical range of alpha acids in the hop.

alphaAcidMax The maximum value for the typical range of alpha acids in the hop.

betaAcidMin The minimum value for the typical range of beta acids in the hop.

betaAcidMax The maxiumum value for the typical range of beta acids in the hop.

humuleneMin The minimum value for the typical range of humulene in the hop. Expressed as a percentage of the total oils.

humuleneMax The maximum value for the typical range of humulene in the hop. Expressed as a percentage of the total oils.

caryophylleneMin The minimum value for the typical range of caryophyllene in the hop. Expressed as a percentage of the total oils.

caryophylleneMax The maximum value for the typical range of caryophyllene in the hop. Expressed as a percentage of the total oils.

cohumuloneMin The minimum value for the typical range of cohumulone in the hop. Expressed as a percentage of the alpha acids.

cohumuloneMax The maximum value for the typical range of cohumulone in the hop. Expressed as a percentage of the alpha acids.

myrceneMin The minimum value for the typical range of myrcene in the hop. Expressed as a percentage of the total oils.

myrceneMax The maximum value for the typical range of myrcene in the hop. Expressed as a percentage of the total oils.

farneseneMin The minimum value for the typical range of farnesene in the hop. Expressed as a percentage of the total oils.

farneseneMax The maximum value for the typical range of farnesene in the hop. Expressed as a percentage of the total oils.

isNoble This is set to Y if the hop is a noble hop.

forBittering Set to Y if the hop is typically used for bittering.

forFlavor Set to Y if the hop is typically used for its flavor.

forAroma Set to Y if the hop is typically used for its aroma.

category This value will always be set to "hop".

categoryDisplay This value will always be set to "Hops".

country This is an object that provides more information about the country matching the hop's countryOfOrigin value.
}
\description{
Gets a list of all hops.  Results are paginated.
}
\concept{Hop}
