% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location.R
\name{getEveryLocation}
\alias{getEveryLocation}
\title{Get All Locations}
\usage{
getEveryLocation(p, ids, locality, region, postalCode, isPrimary,
  inPlanning, isClosed, locationType, countryIsoCode, since, status, order,
  sort)
}
\arguments{
\item{p}{Page Number}

\item{ids}{ID's of the locations to return, comma separated. Max 10.}

\item{locality}{Locality of the location (in US, city)}

\item{region}{Region of the location (in US, state)}

\item{postalCode}{Postal Code of a location}

\item{isPrimary}{Whether the location is the primary one or not}

\item{inPlanning}{Whether the location is in planning and not open yet}

\item{isClosed}{Whether the location is closed.}

\item{locationType}{Key(s) of the type of location. Comma separated. micro macro nano prewpub production office tasting restaurant cidery meadery}

\item{countryIsoCode}{Two-letter country code of the location}

\item{since}{Returns everything that has been updated since that date. Max 30 days. In UNIX timestamp format.}

\item{status}{Status of the location in the API}

\item{order}{How the results should be ordered name Default breweryName locality region postalCode isPrimary inPlanning isClosed locationType countryIsoCode status createDate updateDate}

\item{sort}{How the results should be sorted. ASC Default DESC}
}
\value{
id The unique id of the location.

name The name of the location. Typically this will be like "Main Brewery" for breweries that have one location.

streetAddress The street address of the location.

extendedAddress The extended part of the address of the location. This is normally the second line and could be something like Suite #100.

locality This is the locality / city of the location.

region This is the region / state of the location.

postalCode The postal code / zip code of the location.

phone The phone number of the location.

website The URL to the website for the location.

hoursOfOperation A combination of the hoursOfOperationExplicit and hoursOfOperationNotes field, put together in a human-readable format.

hoursOfOperationExplicitString An array of days and open times. Days may have multiple start and end time sets.

hoursOfOperationNotes A free-text field for notes about the hours of operation.

tourInfo A free-text field containing information about brewery tours at the location.

timezoneId The timezone ID for the location. Example: America/New_York

latitude The latitude of the location.

longitude The longitude of the location.

isPrimary Whether or not this location is the primary location for the associated brewery.

inPlanning Whether or not the location is in planning.

isClosed Whether or not the location is closed (this means permanently closed)

openToPublic Whether or not the location is open to the public.

locationType The key for the type of location. See the locationTypeDisplay for the full display string.

locationTypeDisplay The display string that corresponds to the locationType.

countryIsoCode The two character country code of the location. See the country field for detailed information about the country.

country Detailed information on the country. This field is an object that contains isoCode, name, displayName, isoThree, numbercode, and urlTitle.

yearOpened The year that the location opened.

breweryId The id of the associated brewery. See the brewer object for all the details on the associated brewery.

brewery All the information for the brewery associated with the location.
}
\description{
Gets a listing of all locations.  One of the following attributes must be set:  locality, postalCode, region
}
\concept{Location}
