% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Menu.R
\name{getEveryMenu}
\alias{getEveryMenu}
\title{Gets a listing of beer styles}
\usage{
getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()

getEveryMenu()
}
\arguments{
\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}

\item{none}{}
}
\value{
none

none

none

none

none

none

none

none

none

none

none

none
}
\description{
Gets a listing of beer styles

Gets a listing of all categories

Gets a listing of all glassware

Gets a listing of all SRM values

Gets a listing of all beer availability values

Gets a listing of all fluid sizes

Gets a listing of all beer temperature values

Gets a listing of all countries

Gets a listing of all ingredients

Gets a listing of all location type values

Gets a listing of all fluid size volumes

Gets a listing of all event type values
}
\concept{Menu}
