% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search.R
\name{getEverySearch}
\alias{getEverySearch}
\title{Search}
\usage{
getEverySearch(p, q, type, withBreweries, withSocialAccounts, withGuilds,
  withLocations, withAlternateNames, withIngredients, status)
}
\arguments{
\item{p}{Page Number}

\item{q}{Required Query String}

\item{type}{Search on a specific type of data. Leave blank to search all data. brewery beer guild event}

\item{withBreweries}{Get results with brewery information included. Y N Default}

\item{withSocialAccounts}{Get brewery results with social account information included. Y N Default}

\item{withGuilds}{Get brewery results with guild information included. Y N Default}

\item{withLocations}{Get brewery results with location information included. Y N Default}

\item{withAlternateNames}{Get brewery results with alternate name information included. Y N Default}

\item{withIngredients}{Get beer results with ingredient information included. Y N Default}

\item{status}{System statuses of beer results to return, comma separated. verified Default update_pending Default delete_pending Default}
}
\value{
none
}
\description{
Search
}
\concept{Search}
