% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchGeopoint.R
\name{getEverySearchGeopoint}
\alias{getEverySearchGeopoint}
\title{Search by Coordinate}
\usage{
getEverySearchGeopoint(lat, lng, radius, unit, withSocialAccounts,
  withGuilds, withAlternateNames)
}
\arguments{
\item{lat}{Required Latitude}

\item{lng}{Required Longitude}

\item{radius}{Radius from point. Defaults to 10 miles.}

\item{unit}{Unit of Measure. Defaults to Miles mi Default km}

\item{withSocialAccounts}{Get location results with social account information included. Y N Default}

\item{withGuilds}{Get location results with guild information included. Y N Default}

\item{withAlternateNames}{Get location results with alternate name information included. Y N Default}
}
\value{
none
}
\description{
Search for breweries near a specified coordinate.
}
\concept{SearchGeopoint}
