% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchStyle.R
\name{getEverySearchStyle}
\alias{getEverySearchStyle}
\title{Style Search}
\usage{
getEverySearchStyle(p, q, withDescriptions)
}
\arguments{
\item{p}{Page Number}

\item{q}{Required Query String}

\item{withDescriptions}{Get results with style descriptions included. Y N Default}
}
\value{
none
}
\description{
Search
}
\concept{SearchStyle}
