% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchUpc.R
\name{getEverySearchUpc}
\alias{getEverySearchUpc}
\title{Search by UPC}
\usage{
getEverySearchUpc(p, code)
}
\arguments{
\item{p}{Page Number}

\item{code}{Required UPC}
}
\value{
none
}
\description{
Search for beers by UPC. Will return one or more beers.
}
\concept{SearchUpc}
