% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Socialsite.R
\name{getEverySocialsite}
\alias{getEverySocialsite}
\title{Get List of Social Sites}
\usage{
getEverySocialsite(p)
}
\arguments{
\item{p}{Page Number}
}
\value{
id The unique id of the social site.

name The name of the social site.

website The URL to the website for the social site.
}
\description{
Gets a list of all social sites  Results are paginated.
}
\concept{Socialsite}
