% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fermentable.R
\name{getFermentable}
\alias{getFermentable}
\title{Get A Fermentable}
\usage{
getFermentable(fermentableId)
}
\arguments{
\item{fermentableId}{The fermentableId}
}
\value{
none
}
\description{
Gets a specific fermentable by ID.
}
\concept{Fermentable}
