% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuildSocialaccount.R
\name{getGuildSocialaccount}
\alias{getGuildSocialaccount}
\title{Get Specific Social Account For Guild}
\usage{
getGuildSocialaccount(socialaccountId, guildId)
}
\arguments{
\item{socialaccountId}{The socialaccountId}

\item{guildId}{The guildId}
}
\value{
none
}
\description{
Gets a specific social account for a specific guild.
}
\concept{GuildSocialaccount}
