% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ingredient.R
\name{getIngredient}
\alias{getIngredient}
\title{Get An Ingredient}
\usage{
getIngredient(ingredientId)
}
\arguments{
\item{ingredientId}{The ingredientId}
}
\value{
none
}
\description{
Gets a specific ingredient by ID.
}
\concept{Ingredient}
