% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location.R
\name{getLocation}
\alias{getLocation}
\title{Get a single location}
\usage{
getLocation(locationId)
}
\arguments{
\item{locationId}{The locationId}
}
\value{
none
}
\description{
Gets a single Location
}
\concept{Location}
