% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Socialsite.R
\name{getSocialsite}
\alias{getSocialsite}
\title{Get A Social Site}
\usage{
getSocialsite(socialsiteId)
}
\arguments{
\item{socialsiteId}{The socialsiteId}
}
\value{
none
}
\description{
Gets a specific social site by ID.
}
\concept{Socialsite}
