% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeerSocialaccount.R
\name{updateBeerSocialaccount}
\alias{updateBeerSocialaccount}
\title{Update a Social Account For Beer}
\usage{
updateBeerSocialaccount(socialaccountId, beerId, socialMediaId, handle)
}
\arguments{
\item{socialaccountId}{The socialaccountId}

\item{beerId}{The beerId}

\item{socialMediaId}{Required ID of an existing social media site}

\item{handle}{Required Handle, or identifying marker, for the specific social media site}
}
\value{
none
}
\description{
Edits an existing social account for the beer.
}
\concept{BeerSocialaccount}
