% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brewery.R
\name{updateBrewery}
\alias{updateBrewery}
\title{Update a new brewery}
\usage{
updateBrewery(breweryId, name, description, website, established,
  mailingListUrl, isOrganic, image)
}
\arguments{
\item{breweryId}{The breweryId}

\item{name}{Name of the brewery}

\item{description}{A little bit about the brewery}

\item{website}{Url of the breweries website}

\item{established}{Year brewery was established. Format YYYY}

\item{mailingListUrl}{Url of the breweries mailing list}

\item{isOrganic}{Y N Default}

\item{image}{Base64 encoded image}
}
\value{
none
}
\description{
Update a brewery.  Brewery will be added with status of update_pending.
}
\concept{Brewery}
