% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrewerySocialaccount.R
\name{updateBrewerySocialaccount}
\alias{updateBrewerySocialaccount}
\title{Update a Social Account For Brewery}
\usage{
updateBrewerySocialaccount(socialaccountId, breweryId, socialmediaId,
  handle)
}
\arguments{
\item{socialaccountId}{The socialaccountId}

\item{breweryId}{The breweryId}

\item{socialmediaId}{Required ID of an existing social media site}

\item{handle}{Required Handle, or identifying marker, for the specific social media site}
}
\value{
none
}
\description{
Edits an existing social account for the brewery.
}
\concept{BrewerySocialaccount}
