% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guild.R
\name{updateGuild}
\alias{updateGuild}
\title{Update an guild}
\usage{
updateGuild(guildId, name, description, established, website, image)
}
\arguments{
\item{guildId}{The guildId}

\item{name}{Required Name of the guild}

\item{description}{Description of the guild}

\item{established}{Year the guild was established}

\item{website}{Website of the guild}

\item{image}{An image or logo for the guild, base64 encoded source.}
}
\value{
none
}
\description{
Updates an existing guild.
}
\concept{Guild}
